Fix hang in exception reporting dialog
ROX-Lib's exception dialog often fails to work correctly. It appears,
but doesn't respond to any events, and neither does the rest of the
application. I don't think it always happens, but I haven't figured out
the circumstances of when it works and when it doesn't.

Note that this reverses a previous change (8c8822a3), so presumably the
code was like this for a reason. However, the reason wasn't documented
so let's just try it and see...